backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
djangorestframework==3.14.0
